% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_soas.R
\name{align_soas}
\alias{align_soas}
\title{aligns SOAs}
\usage{
align_soas(
  soa_vec,
  n_soa = 20,
  round_factor = 2,
  soa_spacing = 42,
  verbose = TRUE,
  difference_plot = FALSE
)
}
\arguments{
\item{soa_vec}{vector of SOAs}

\item{n_soa}{number of SOAs in the paradigm}

\item{round_factor}{defines the second argument for the round() command decimal value with which to start the rounding procedure.}

\item{soa_spacing}{needs ms value of spacing between SOAs. Will use this to group all SOAs with a difference < soa_spacing. By default, it will try to achieve n_soa SOAs with round_factor =2 and will give out warnings.}

\item{verbose}{def = T}

\item{difference_plot}{def = F}
}
\description{
\code{align_soas}
evaluates a vector of SOAs and tries to align slightly different SOA values by rounding them in a way that we end up in n_soa different SOAs.
If you want, align_soas can plot a histogram of difference between original input vector and the output vector provided by this function.
}
\examples{
repeated_data = rep_dataframe(data, 2, "rbind")
}
\author{
René Michel
}
