% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{BiasCEsts}
\alias{BiasCEsts}
\title{Performs BiasCEsts}
\usage{
BiasCEsts(circdat, indsym, n)
}
\description{
\code{BiasCEsts}
Performs BiasCEsts with code from (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
Rene Michel
}
