% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_div.R
\name{check_div}
\alias{check_div}
\title{check_div}
\arguments{
\item{vec}{vector of numbers to be tested}

\item{num}{number to be divided}
}
\description{
\code{check_div}
returns all elements of a vector that produce a modulo of 0 for a given number.
}
\examples{
check_div(seq(2,256,2), 456)
}
\author{
Ren攼㸹 Michel
}
