% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{CircFisherTest}
\alias{CircFisherTest}
\title{Performs circular Fisher Test}
\usage{
CircFisherTest(cdat, ndat, g, parametric = T, NR = 0)
}
\arguments{
\item{cdat}{combined vector of observations, e.g. c(Obs1_1,...Obs1_10, Obs2_1, ... Obs2_20)}

\item{ndat}{vector with N observations of each group, e.g. c(10,20) for one group n = 10 and another of n = 20}

\item{g}{number of groups}

\item{parametric}{defaults to T}

\item{NR}{defaults to 0, number of permutations}
}
\description{
\code{CircFisherTest}
Performs circular Fisher Test (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
René Michel
}
