% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_value.R
\name{closest_value}
\alias{closest_value}
\title{closest_value}
\arguments{
\item{vec1}{vector of numbers to find neighbours for}

\item{vec2}{vector of numbers to find neighbours from (e.g. the new bins)}
}
\description{
\code{closest_value}
returns the numerically closest value in vector 2 for each element of vector 1
}
\examples{
a = seq(1,100,2)
b = seq(10,100,10)
closest_value(a,b)
}
\author{
Ren攼㸹 Michel
}
