% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{ConfIntBoot}
\alias{ConfIntBoot}
\title{Performs ConfIntBoot}
\usage{
ConfIntBoot(origdat, indsym, conflevel, B)
}
\description{
\code{ConfIntBoot}
Performs ConfIntBoot with code from (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
Rene Michel
}
