% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{ConfIntLS}
\alias{ConfIntLS}
\title{Performs ConfIntLS}
\usage{
ConfIntLS(circdat, indsym, conflevel)
}
\description{
\code{ConfIntLS}
Performs ConfIntLS with code from (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
Rene Michel
}
