% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.data.frame.R
\name{create.data.frame}
\alias{create.data.frame}
\title{create dataframe}
\arguments{
\item{default_value}{which value to add in all cells? defaults to NA}

\item{nrow}{number of rows}

\item{ncol}{number of cols}

\item{col.names}{character vector with colnames}
}
\description{
\code{create.data.frame}
creates a data.frame and allows to add colnames
}
\examples{
d = create.data.frame(NA, 2, 2, c("Col1", "Col2"))
}
\author{
Ren攼㸹 Michel
}
