% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_dirs}
\alias{create_dirs}
\title{Creates a number of directories}
\usage{
create_dirs(...)
}
\arguments{
\item{...}{requires a vector of character strings.}
}
\description{
\code{create_dirs}
Creates specified folders in subsequent order.
Is able to create nested folders - for this, the outer folder must be parsed before the inner ones.
}
\examples{
# create 2 new folders
dirs = c("Testfolder", "Testfolder2")
create_dirs(dirs)
}
\author{
René Michel
}
