% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFacetWrap.R
\docType{data}
\name{CustomFacetWrap}
\alias{CustomFacetWrap}
\title{ggproto CFW}
\format{
An object of class \code{CustomFacetWrap} (inherits from \code{FacetWrap}, \code{Facet}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
CustomFacetWrap
}
\description{
\code{CustomFacetWrap}
ggproto for facet_wrap_custom
}
\author{
René Michel
}
\keyword{datasets}
