% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_fft.R
\name{dsa_fft}
\alias{dsa_fft}
\title{performs an FFT on behavioral data}
\usage{
dsa_fft(
  data,
  time = "col",
  detrend = "linear",
  window = "tukey",
  scaling = "z",
  padding = "zeros",
  desired_res = 1,
  sfreq = 25,
  desired_out = "amp",
  single_sided = T,
  out = "mod",
  alpha = 0.1,
  verbose = T
)
}
\arguments{
\item{data}{requires the dataframe}

\item{time}{are timepoints across columns or rows? default = "col"}

\item{detrend}{Detrending the data (default = 'linear', also available: 'dc', '2ndorder' or 'none')}

\item{window}{Windowing whole time series (default = 'tukey', also available: 'hanning', 'none')}

\item{scaling}{Rescaling the data (default: 'z' for z-transform, also available: 'none' )}

\item{padding}{Padding the data (default: 'zeros' for zero-padding, also available: 'mean', 'none')}

\item{desired_res}{Padding the data up to desired_res (automatically evaluates necessary number of pads to achieve the resolution, default = 1)}

\item{sfreq}{sampling frequency of inserted datapoints}

\item{desired_out}{Extract amplitude values (default: 'amp', also available 'power' or 'logPower')}

\item{single_sided}{Single-sided frequency spectum (default = T)}

\item{out}{By default, output is Mod (Amplitude) of the signal. Change to 'arg' to get phase or 'complex' to get the complex output.}

\item{alpha}{only for tukey window}

\item{verbose}{T/F}
}
\description{
\code{dsa_fft}
A function to compute the fft results for a time series (especially dense sampled behavioral data).
By default, dsa_fft will treat the data so that each column is considered one timepoint and computes all steps by row.
If your input data is organized with time = rows, change time argument to 'row'.
Both sampling frequency and desired frequency resolution (Bin width per frequency bin of FFT) needs to be in Hz.
}
\author{
Rene Michel
}
