% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{dValues}
\alias{dValues}
\title{Computes dValues}
\usage{
dValues(cdat, ndat, g)
}
\arguments{
\item{ndat}{vector with N observations of each group, e.g. c(10,20) for one group n = 10 and another of n = 20}

\item{g}{number of groups}

\item{dvals}{dvals}
}
\description{
\code{dValues}
Computes dValues (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
René Michel
}
