% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac2num.R
\name{fac2num}
\alias{fac2num}
\title{convert factor to numeric}
\usage{
fac2num(c)
}
\arguments{
\item{c}{the factor to be converted}
}
\description{
\code{fac2num}
converts factor levels to numeric values
}
\author{
René Michel
}
