% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_wrap_costum.R
\name{facet_wrap_custom}
\alias{facet_wrap_custom}
\title{customize facet scales}
\usage{
facet_wrap_custom(..., scale_overrides = NULL)
}
\arguments{
\item{scales_overrides}{with arg which (specifies x/y axis, e.g. 1 / 2) and scale}
}
\description{
\code{facet_wrap_custom}
ggplots facet_wrap with an extra argument "scales_overrides"
}
\author{
René Michel
}
