% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{FgVal}
\alias{FgVal}
\title{Computes FgVal}
\usage{
FgVal(dvals, ndat, g)
}
\arguments{
\item{dvals}{dvals}

\item{ndat}{vector with N observations of each group, e.g. c(10,20) for one group n = 10 and another of n = 20}

\item{g}{number of groups}
}
\description{
\code{MooreRTestRand}
Computes FgVal (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
René Michel
}
