% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitViolins.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{introduce geom_split_violin}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
\code{geom_split_violin}
Plot violins by a split (e.g. by validity) with typical mapping specs
}
\author{
René Michel
}
