% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitViolins.R
\docType{data}
\name{GeomSplitViolin}
\alias{GeomSplitViolin}
\title{enables geom_split_violin}
\format{
An object of class \code{GeomSplitViolin} (inherits from \code{GeomViolin}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomSplitViolin
}
\description{
\code{GeomSplitViolin}
enables geom_split_violin
}
\author{
René Michel
}
\keyword{datasets}
