% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{Loads an R file and returns it}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{the filename}
}
\description{
\code{loadRData}
Loads an R file and returns it (c) ricardo
}
\author{
René Michel
}
