% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{MooreRStats}
\alias{MooreRStats}
\title{Computes required stats for a Moore Test for circular data}
\usage{
MooreRStats(ldat1, ldat2)
}
\arguments{
\item{ldat1}{vec 1}

\item{ldat2}{vec 2}
}
\description{
\code{MooreRStats}
Computes required stats for a Moore Test for circular data (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
René Michel
}
