% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{MooreRTestRand}
\alias{MooreRTestRand}
\title{Computes Moore Test}
\usage{
MooreRTestRand(cosphi, sinphi, Ranks, NR)
}
\arguments{
\item{cosphi}{cosphi}

\item{sinphi}{sinphi}

\item{Ranks}{ranks}

\item{NR}{number of permutations}
}
\description{
\code{MooreRTestRand}
Computes Moore Test for circular data (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
René Michel
}
