% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{MooreRTestStat}
\alias{MooreRTestStat}
\title{Computes Moore Test statistic}
\usage{
MooreRTestStat(cosphi, sinphi, Ranks)
}
\arguments{
\item{cosphi}{cosphi}

\item{sinphi}{sinphi}

\item{Ranks}{ranks}
}
\description{
\code{MooreRTestStat}
Computes Moore Test statistic for circular data with code from (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
René Michel
}
