% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveme.R
\name{moveme}
\alias{moveme}
\title{Moves a column in a dataframe to a specified position.}
\usage{
moveme(invec, movecommand)
}
\arguments{
\item{invec}{vector of names of the dataframe of interest gathered by names(dataframe)}

\item{movecommand}{character string vector containing one of four movements "before", "after", "first" or "last"}
}
\description{
\code{moveme}
Moves a column in a dataframe to a specified position.
}
\examples{
# move the column to be the first column in the dataframe
data = data[moveme(names(data),'var first')]

}
\author{
René Michel
}
