% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_hist.R
\name{plot_gg_hist}
\alias{plot_gg_hist}
\title{creates a predefined histogram}
\usage{
plot_gg_hist(
  data,
  xvar = "DummyVar",
  plot_title = "Dummy Title",
  facets = NA,
  fill = NA,
  binwidth = 10,
  col = "black",
  fillcol = "grey",
  breaks = seq(-180, 180, 20)
)
}
\arguments{
\item{data}{data}

\item{xvar}{xvar}

\item{plot_title}{title}

\item{facets}{facet var}

\item{fill}{fill?}

\item{binwidth}{bin width}

\item{col}{line col}

\item{fillcol}{fill col}

\item{breaks}{breaks}
}
\description{
\code{plot_gg_hist}
Once called, it simply enables us to add + theme_RM to ggplots
}
\author{
René Michel
}
