% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_dataframe.R
\name{rep_dataframe}
\alias{rep_dataframe}
\title{Creates a number of directories}
\usage{
rep_dataframe(data, times = 1, by = "rbind")
}
\arguments{
\item{data}{requires the dataframe to be copied.}

\item{times}{requires the number of times the dataframe has to be repeated. Defaults to 1.}

\item{by}{requires a character string specifying the function to bind the repeated dataframes("rbind", "cbind"). Defaults to "rbind".}
}
\description{
\code{rep_dataframe}
Repeats a dataframe n times by rbind or cbind.
}
\examples{
repeated_data = rep_dataframe(data, 2, "rbind")
}
\author{
René Michel
}
