% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_downsample.R
\name{rm_downsample}
\alias{rm_downsample}
\title{rm_downsample}
\arguments{
\item{vec}{data vector}

\item{t}{time vector}

\item{f}{factor of downsampling (integer)}

\item{method}{"filter" or "average" (uses time windows of f to average the time course) or "rollmean" (from zoo package)}
}
\description{
\code{rm_downsample}
downsample a data vector and transform its corresponding time vector
}
\examples{
rm_downsample(c(1,-1,3, 4), c(1,2,3,4), 2)
}
\author{
René Michel
}
