% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_install_packages.R
\name{rm_install_packages}
\alias{rm_install_packages}
\title{rm_install_packages}
\usage{
rm_install_packages(
  git_pkgs = NULL,
  git_private_pkgs = NULL,
  cran_pkgs = NULL,
  load_after_install = T
)
}
\arguments{
\item{git_pkgs}{A character vector with all packages from public github repositories, specified as "username/reponame@branch"}

\item{git_private_pkgs}{A character vector with all packages from private github repositories, specified as "username/reponame@branch"}

\item{cran_pkgs}{A character vector with all package names from cran}

\item{load_after_install}{defaults to T, loads all pkgs after installation}
}
\value{

}
\description{
rm_install_packages
}
