% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{RSTestBoot}
\alias{RSTestBoot}
\title{Performs RSTestBoot}
\usage{
RSTestBoot(origdat, B)
}
\description{
\code{RSTestBoot}
Performs RSTestBoot with code from (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
Rene Michel
}
