% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{RSTestStat}
\alias{RSTestStat}
\title{Performs RRSTestStat}
\usage{
RSTestStat(circdat)
}
\description{
\code{RSTestStat}
Performs RSTestStat with code from (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
Rene Michel
}
