% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_override.R
\name{scale_override}
\alias{scale_override}
\title{specify which scale to override}
\usage{
scale_override(which, scale)
}
\description{
\code{scale_override}
subfunction for facet_wrap_custom
}
\author{
René Michel
}
