% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_repo_path.R
\name{set_repo_path}
\alias{set_repo_path}
\title{set_repo_path}
\usage{
set_repo_path(...)
}
\arguments{
\item{...}{(Multiple) vectors of string pairs, where the first string is always the computername and the second string the desired repo path for that pc}
}
\value{
character
}
\description{
set_repo_path
}
