% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theme_RM.R
\name{set_theme_RM}
\alias{set_theme_RM}
\title{enables RM_theme for ggplot2}
\usage{
set_theme_RM()
}
\description{
\code{set_theme_RM}
Once called, it simply enables us to add + theme_RM to ggplots
}
\examples{

set_theme_RM()
ggplot(...)+theme_RM

}
\author{
René Michel
}
