% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro.by.R
\name{shapiro.by}
\alias{shapiro.by}
\title{Perform shapiro test by group}
\arguments{
\item{data}{vector or dataframe with multiple columns which contain the data}

\item{group}{group variable which is used to split the data vector(s)}
}
\description{
\code{shapiro.by}
Adopted from Marco Sandri.
}
\examples{
res = shapiro.by(data$reactiontime, data$group)
}
\author{
René Michel
}
