% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiroby.R
\name{shapiroby}
\alias{shapiroby}
\title{Perform shapiro test by group}
\usage{
shapiroby(data, group)
}
\arguments{
\item{data}{vector or dataframe with multiple columns which contain the data}

\item{group}{group variable which is used to split the data vector(s)}
}
\description{
\code{shapiroby}
Adopted from Marco Sandri.
}
\examples{
res = shapiro.by(data$reactiontime, data$group)
}
\author{
René Michel
}
