% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{SpecMeanTestBoot}
\alias{SpecMeanTestBoot}
\title{Performs SpecMeanTestBoot}
\usage{
SpecMeanTestBoot(origdat, mu0, indsym, B)
}
\description{
\code{SpecMeanTestBoot}
Performs SpecMeanTestBoot with code from (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
Rene Michel
}
