% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryRMwithin.R
\name{summaryRMwithin}
\alias{summaryRMwithin}
\title{summaryRMwithin with extras}
\usage{
summaryRMwithin(
  data = NULL,
  measurevar,
  betweenvars = NULL,
  withinvars = NULL,
  idvar = NULL,
  doNorm = T,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{doNorm}{(default = T)}
}
\description{
\code{summaryRMwithin}
Adds option "doNorm" to summarySEwithin, to have control over normalization of means. Note: Correction is still applied using normalized data (to get adequate estimates for SE), but means stay unaffected
}
\author{
René Michel
}
