% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellmeFFT.R
\name{tellmeFFT}
\alias{tellmeFFT}
\title{tellmeFFT}
\usage{
tellmeFFT(
  spaceBetweenIntervals = NA,
  numSoa = NA,
  startingSoa = 0,
  freqOfInterest = 4
)
}
\arguments{
\item{spaceBetweenIntervals}{Spacing between SOAs in ms}

\item{numSoa}{number of SOAs in the paradigm}

\item{startingSoa}{first SOA in ms after time locking event (e.g. cue)}

\item{freqOfInterest}{which frequency are you interested? in Hz}
}
\description{
\code{tellmeFFT}
informs you about the FFT characteristics for your proposed dense sampling study.
}
\examples{
tellmeFFT(42,20,192,4)
}
\author{
René Michel
}
