% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTestFunctions.R
\name{WallraffTest}
\alias{WallraffTest}
\title{Performs a Wallraff Test for circular data}
\usage{
WallraffTest(cdat, ndat, g)
}
\arguments{
\item{cdat}{combined vector of observations, e.g. c(Obs1_1,...Obs1_10, Obs2_1, ... Obs2_20)}

\item{ndat}{vector with N observations of each group, e.g. c(10,20) for one group n = 10 and another of n = 20}

\item{g}{number of groups}
}
\description{
\code{WallraffTest}
Performs a Wallraff Test for circular data (from Pewsey, A., Neuhäuser, M., & Ruxton, G. D. (2013). Circular statistics in R. Oxford University Press.)
}
\author{
René Michel
}
